Move value objects to vo folder
