+++ src/modules/files-storage/domain/factory/wopi-access-token.factory.ts
@@ -1,4 +1,4 @@
-import { WopiAccessToken } from '../wopi-access-token.vo';
+import { WopiAccessToken } from '../vo';
 
 export class WopiAccessTokenFactory {
 	public static build(props: WopiAccessToken): WopiAccessToken {
