+++ src/modules/files-storage/domain/file-record.factory.ts
@@ -1,7 +1,7 @@
 import { ObjectId } from '@mikro-orm/mongodb';
 import { EntityId } from '@shared/domain/types';
 import { FileRecord, FileRecordProps, ParentInfo } from './file-record.do';
-import { FileRecordSecurityCheck } from './security-check.vo';
+import { FileRecordSecurityCheck } from './vo';
 
 export class FileRecordFactory {
 	private static build(fileRecordProps: FileRecordProps, securityCheck: FileRecordSecurityCheck): FileRecord {
