+++ src/modules/files-storage/domain/index.ts
@@ -4,8 +4,13 @@ export * from './factory';
 export * from './file-record.do';
 export * from './interface';
 export { ParentStatisticFactory } from './parent-statistic.factory';
-export { ParentStatistic, ParentStatisticProps } from './parent-statistic.vo';
-export { FileRecordSecurityCheck, FileRecordSecurityCheckProps, ScanStatus } from './security-check.vo';
 export { FilesStorageService, PreviewService } from './service';
-export * from './wopi-access-token.vo';
-export * from './wopi-payload.vo';
+export {
+	FileRecordSecurityCheck,
+	FileRecordSecurityCheckProps,
+	ParentStatistic,
+	ParentStatisticProps,
+	ScanStatus,
+} from './vo';
+export * from './vo/wopi-access-token.vo';
+export * from './vo/wopi-payload.vo';
