+++ src/modules/files-storage/domain/service/files-storage-delete.service.spec.ts
@@ -11,7 +11,7 @@ import { FILES_STORAGE_S3_CONNECTION, FileStorageConfig } from '../../files-stor
 import { fileRecordTestFactory } from '../../testing';
 import { FileRecord, FileRecordProps } from '../file-record.do';
 import { FILE_RECORD_REPO, FileRecordRepo, StorageLocation } from '../interface';
-import { FileRecordSecurityCheckProps } from '../security-check.vo';
+import { FileRecordSecurityCheckProps } from '../vo';
 import { FilesStorageService } from './files-storage.service';
 
 describe('FilesStorageService delete methods', () => {
