+++ src/modules/files-storage/domain/service/files-storage.service.ts
@@ -20,8 +20,8 @@ import { FileRecordFactory } from '../file-record.factory';
 import { CopyFileResult, FILE_RECORD_REPO, FileRecordRepo, GetFileResponse, StorageLocationParams } from '../interface';
 import { FileStorageActionsLoggable } from '../loggable';
 import { FileResponseBuilder, ScanResultDtoMapper } from '../mapper';
-import { ParentStatistic } from '../parent-statistic.vo';
-import { ScanStatus } from '../security-check.vo';
+
+import { ParentStatistic, ScanStatus } from '../vo';
 import { ArchiveFactory } from './archive.factory';
 import { fileTypeStream } from './file-type.helper';
