+++ src/modules/files-storage/domain/vo/index.ts
@@ -1,0 +1,4 @@
+export * from './parent-statistic.vo';
+export * from './security-check.vo';
+export * from './wopi-access-token.vo';
+export * from './wopi-payload.vo';
