+++ src/modules/files-storage/domain/vo/parent-statistic.vo.ts
