+++ src/modules/files-storage/domain/vo/security-check.vo.spec.ts
