+++ src/modules/files-storage/domain/vo/wopi-access-token.vo.spec.ts
