+++ src/modules/files-storage/domain/vo/wopi-payload.vo.ts
